requirements.txt
keras
tensorflow
numpy
Pillow